Add files via upload
Note: This tool requires python version 3 or newer.

SETUP.py:
This python file will create an appropriate environment for the NetInspector to run error-free. Run this file on your system before getting started with scanning and analyzing packets using NetInspector.

NetInspector.py:
Tool with an ability of port scanner and packet analyzer with interactive environment and easy to understand flow of operations.